refactor: reorganize CLI commands into subdirectory structure (#183)
- Move workspace, config, and hooks commands into dedicated directories
- Split large command files into individual subcommand files
- Update imports and package structure for consistency
- Fix test failures by using direct function calls instead of rootCmd
- Align with session command directory pattern

This provides better code organization and makes the codebase more maintainable
by grouping related commands together in their own packages.
