+++ internal/cli/commands/config/config.go
@@ -1,9 +1,11 @@
-// Package commands provides CLI command implementations for amux.
-package commands
+// Package config provides CLI command implementations for amux config management.
+package config
 
 import (
 	"github.com/spf13/cobra"
 )
+
+var showFormat string
 
 var configCmd = &cobra.Command{
 	Use:   "config",
@@ -22,5 +24,16 @@ The config command provides subcommands to view and edit the configuration file.
 }
 
 func init() {
-	rootCmd.AddCommand(configCmd)
+	// Add subcommands
+	configCmd.AddCommand(configShowCmd)
+	configCmd.AddCommand(configEditCmd)
+	configCmd.AddCommand(configValidateCmd())
+
+	// Configure flags
+	configShowCmd.Flags().StringVar(&showFormat, "format", "yaml", "Output format (yaml, json, pretty)")
+}
+
+// Command returns the config command
+func Command() *cobra.Command {
+	return configCmd
 }
