+++ internal/cli/commands/config/config_validate_test.go
@@ -1,4 +1,4 @@
-package commands
+package config
 
 import (
 	"bytes"
