+++ internal/cli/commands/workspace/helpers.go
@@ -1,0 +1,27 @@
+package workspace
+
+import (
+	"fmt"
+
+	"github.com/aki/amux/internal/core/config"
+	"github.com/aki/amux/internal/core/workspace"
+)
+
+func getWorkspaceManager() (*workspace.Manager, error) {
+	// Find project root
+	projectRoot, err := config.FindProjectRoot()
+	if err != nil {
+		return nil, err
+	}
+
+	// Create configuration manager
+	configManager := config.NewManager(projectRoot)
+
+	// Ensure initialized
+	if !configManager.IsInitialized() {
+		return nil, fmt.Errorf("amux not initialized. Run 'amux init' first")
+	}
+
+	// Create workspace manager
+	return workspace.NewManager(configManager)
+}
