+++ internal/cli/commands/workspace/workspace_test.go
@@ -1,4 +1,4 @@
-package commands
+package workspace
 
 import (
 	"context"
